Conflicts:
	backend/src/main/java/cz/cuni/xrg/intlib/backend/execution/PipelineWorker.java
	commons-app/src/main/java/cz/cuni/xrg/intlib/commons/app/execution/ExecutionContextImpl.java
	frontend/src/main/java/cz/cuni/xrg/intlib/frontend/gui/components/DebuggingView.java
	frontend/src/main/java/cz/cuni/xrg/intlib/frontend/gui/components/QueryView.java

+++ commons-app/src/main/java/cz/cuni/xrg/intlib/commons/app/pipeline/PipelineFacade.java
@@ -135,7 +135,7 @@ public class PipelineFacade {
 	 * @param id of PipelineExecution
 	 * @return PipelineExecution
 	 */
-	public PipelineExecution getExecution(int id) {
+	public PipelineExecution getExecution(long id) {
 		return em.find(PipelineExecution.class, id);
 	}
